+++ src/Html2OpenXml/Expressions/Numbering/HeadingElementExpression.cs
@@ -9,6 +9,7 @@
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A
  * PARTICULAR PURPOSE.
  */
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -23,7 +24,8 @@ namespace HtmlToOpenXml.Expressions;
 /// </summary>
 sealed class HeadingElementExpression(IHtmlElement node) : NumberingExpressionBase(node)
 {
-    private static readonly Regex numberingRegex = new(@"^\s*(\d+\.?)*\s*");
+    private static readonly Regex numberingRegex = new(@"^\s*(?<number>[0-9\.]+\s*)[^0-9]",
+        RegexOptions.Compiled, TimeSpan.FromMilliseconds(100));
 
     /// <inheritdoc/>
     public override IEnumerable<OpenXmlElement> Interpret (ParsingContext context)
@@ -36,7 +38,7 @@ sealed class HeadingElementExpression(IHtmlElement node) : NumberingExpressionBa
 
         var paragraph = childElements.FirstOrDefault() as Paragraph;
 
-        paragraph ??= new Paragraph(childElements);
+        paragraph ??= new(childElements);
         paragraph.ParagraphProperties ??= new();
         paragraph.ParagraphProperties.ParagraphStyleId = 
             context.DocumentStyle.GetParagraphStyle(context.DocumentStyle.DefaultStyles.HeadingStyle + level);
@@ -65,16 +67,30 @@ sealed class HeadingElementExpression(IHtmlElement node) : NumberingExpressionBa
 
     private static bool IsNumbering(OpenXmlElement runElement)
     {
+        if (runElement.InnerText is null)
+            return false;
+
         // Check if the line starts with a number format (1., 1.1., 1.1.1.)
         // If it does, make sure we make the heading a numbered item
-        Match regexMatch = numberingRegex.Match(runElement.InnerText ?? string.Empty);
+        var headingText = runElement.InnerText;
+        Match regexMatch;
+        try
+        {
+            regexMatch = numberingRegex.Match(headingText);
+        }
+        catch (RegexMatchTimeoutException)
+        {
+            return false;
+        }
+
 
         // Make sure we only grab the heading if it starts with a number
-        if (regexMatch.Groups.Count > 1 && regexMatch.Groups[1].Captures.Count > 0)
+        if (regexMatch.Success && headingText.Length > regexMatch.Groups["number"].Length)
         {
-             // Strip numbers from text
+            // Strip numbers from text
+            headingText = headingText.Substring(regexMatch.Groups["number"].Length);
             runElement.InnerXml = runElement.InnerXml
-                .Replace(runElement.InnerText!, runElement.InnerText!.Substring(regexMatch.Length));
+                .Replace(runElement.InnerText!, headingText);
 
             return true;
         }