+++ src/Html2OpenXml/Expressions/Table/TableExpression.cs
@@ -156,7 +156,7 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
                 }
             }
 
-            if (rows.Any())
+            if (rows.Length > 0)
                 columnCount = Math.Max(rows.Max(), columnCount);
         }
 
@@ -186,6 +186,9 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
                 tableProperties.TableWidth = new() { Type = TableWidthUnitValues.Dxa, 
                     Width = width.ValueInDxa.ToString(CultureInfo.InvariantCulture) };
                 break;
+            case UnitMetric.Auto:
+                tableProperties.TableWidth = new() { Width = "0", Type = TableWidthUnitValues.Auto };
+                break;
         }
 
         foreach (string className in tableNode.ClassList)
@@ -197,10 +200,6 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
                 break;
             }
         }
-
-        var align = Converter.ToParagraphAlign(tableNode.GetAttribute("align"));
-        if (align.HasValue)
-            tableProperties.TableJustification = new() { Val = align.Value.ToTableRowAlignment() };
 
         var dir = tableNode.GetTextDirection();
         if (dir.HasValue)
@@ -243,7 +242,8 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
             tableProperties.TableBorders = tableBorders;
         }
         // is the border=0? If so, we remove the border regardless the style in use
-        else if (tableNode.Border == 0)
+        // but only remove border if the html style border was set, otherwise leave the border style as-is.
+        else if (!styleBorder.IsEmpty && tableNode.Border == 0)
         {
             tableProperties.TableBorders = new TableBorders() {
                 TopBorder = new TopBorder { Val = BorderValues.None },
@@ -280,5 +280,22 @@ sealed class TableExpression(IHtmlTableElement node) : PhrasingElementExpression
                 };
             }
         }
+
+        var align = Converter.ToParagraphAlign(tableNode.GetAttribute("align"))
+            ?? Converter.ToParagraphAlign(styleAttributes["justify-self"]);
+        if (!align.HasValue)
+        {
+            var margin = styleAttributes.GetMargin("margin");
+            if (margin.Left.Type == UnitMetric.Auto)
+            {
+                if (margin.Right.Type == UnitMetric.Auto)
+                    align = JustificationValues.Center;
+                else
+                    align = JustificationValues.Right;
+            }
+        }
+
+        if (align.HasValue)
+            tableProperties.TableJustification = new() { Val = align.Value.ToTableRowAlignment() };
     }
 }