+++ test/HtmlToOpenXml.Tests/TableTests.cs
@@ -247,7 +247,7 @@ namespace HtmlToOpenXml.Tests
         }
 
         [TestCaseSource(nameof(BorderWidthCases))]
-        public void HtmlBorders_ShouldSucceed(string borderAtrribute, IEnumerable<string> expectedBorderValue, IEnumerable<uint?> expectedBorderWidth)
+        public void HtmlBorders_ShouldSucceed(string borderAtrribute, IEnumerable<string>? expectedBorderValue, IEnumerable<uint?>? expectedBorderWidth)
         {
             // we specify a style which doesn't handle borders
             converter.HtmlStyles.AddStyle(new Style {
@@ -260,6 +260,12 @@ namespace HtmlToOpenXml.Tests
             Assert.That(elements, Has.Count.EqualTo(1));
             Assert.That(elements, Has.All.TypeOf<Table>());
             var borders = elements[0].GetFirstChild<TableProperties>()?.TableBorders;
+            if (expectedBorderValue is null)
+            {
+                Assert.That(borders, Is.Null);
+                return;
+            }
+
             Assert.That(borders, Is.Not.Null);
             Assert.That(borders.HasChild<BorderType>(), Is.True);
             Assert.That(new string?[] { borders.TopBorder?.Val?.InnerText,
@@ -289,8 +295,8 @@ namespace HtmlToOpenXml.Tests
         static readonly object[] BorderWidthCases =
         [
             // Negative border should be considered as zero
-            new object[] { "border='-1'", Enumerable.Repeat("none", 6), null! },
-            new object[] { "border='0'", Enumerable.Repeat("none", 6), null! },
+            new object[] { "border='-1'", null!, null! },
+            new object[] { "border='0'", null!, null! },
             new object[] { "border='1'",
                 new string[] { "none", "none", "none", "none", "single", "single" }, 
                 new uint?[] { null, null, null, null, 14, 14 } },
@@ -330,9 +336,9 @@ namespace HtmlToOpenXml.Tests
             });
         }
 
-        [TestCase("right", "right")]
-        [TestCase("", "center")]
-        public void TableCaptionAlign_ReturnsPositionedParagraph_AlignedWithTable(string alignment, string expectedAlign)
+        [TestCase("right", ExpectedResult = "right")]
+        [TestCase("", ExpectedResult = "center")]
+        public string? TableCaptionAlign_ReturnsPositionedParagraph_AlignedWithTable(string alignment)
         {
             var elements = converter.Parse(@$"<table align=""center"">
                     <caption align=""{alignment}"">Some table caption</caption>
@@ -341,7 +347,35 @@ namespace HtmlToOpenXml.Tests
 
             Assert.That(elements, Has.Count.EqualTo(2));
             var caption = (Paragraph) elements[1];
-            Assert.That(caption.ParagraphProperties?.Justification?.Val?.ToString(), Is.EqualTo(expectedAlign));
+            return caption.ParagraphProperties?.Justification?.Val?.ToString();
+        }
+
+        [TestCase("align='right'", ExpectedResult = "right")]
+        [TestCase("style='justify-self:center'", ExpectedResult = "center")]
+        [TestCase("style='margin-left:auto'", ExpectedResult = "right")]
+        [TestCase("style='margin-left:auto;margin-right:auto'", ExpectedResult = "center")]
+        public string? TableAlign_ReturnsTableJustification(string style)
+        {
+            var elements = converter.Parse(@$"<table {style}>
+                    <tr><td>Cell 1.1</td></tr>
+                </table>");
+
+            Assert.That(elements, Has.Count.EqualTo(1));
+            return elements[0].GetFirstChild<TableProperties>()?.TableJustification?.Val?.ToString();
+        }
+
+        [TestCase("", ExpectedResult = "right")]
+        [TestCase("justify-self:center", ExpectedResult = "center")]
+        public string? NestedTableAlign_ReturnsTableOrParentJustification(string tableStyle)
+        {
+            var elements = converter.Parse(@$"<div style='justify-content: right'>
+                <table style='{tableStyle}'>
+                    <tr><td>Cell 1.1</td></tr>
+                </table>
+                </div>");
+
+            Assert.That(elements, Has.Count.EqualTo(1));
+            return elements[0].GetFirstChild<TableProperties>()?.TableJustification?.Val?.ToString();
         }
 
         [Test]
@@ -467,6 +501,7 @@ namespace HtmlToOpenXml.Tests
             var cell = elements[0].GetFirstChild<TableRow>()?.GetFirstChild<TableCell>();
             Assert.That(cell, Is.Not.Null);
             Assert.That(cell.HasChild<Table>(), Is.True);
+            Assert.That(cell.HasChild<Paragraph>(), Is.True, "Word requires at least a paragraph");
         }
 
         [Test]
@@ -626,6 +661,27 @@ namespace HtmlToOpenXml.Tests
             var runs = paragraphs.First().Elements<Run>();
             Assert.That(runs.Count(), Is.EqualTo(1));
             Assert.That(runs.First().RunProperties?.Border, Is.Null);
+        }
+
+        [TestCase("100%", "pct", "5000")]
+        [TestCase("auto", "auto", "0")]
+        [TestCase("120px", "dxa", "1800")]
+        [TestCase("", "pct", "5000", Description = "Defaults to 100%")]
+        public void Width_ReturnsRefineTableWidth(string width, string expectedUnit, string expectedValue)
+        {
+            var elements = converter.Parse(@$"<table style='width: {width}'>
+                <tr><td>Placeholder</td></tr>
+            </table>");
+
+            Assert.That(elements, Has.Count.EqualTo(1));
+            Assert.That(elements, Has.All.TypeOf<Table>());
+            var tableWidth = elements[0].GetFirstChild<TableProperties>()?.TableWidth;
+            Assert.That(tableWidth, Is.Not.Null);
+            Assert.Multiple(() =>
+            {
+                Assert.That(tableWidth?.Type?.Value, Is.EqualTo(new TableWidthUnitValues(expectedUnit)));
+                Assert.That(tableWidth?.Width?.Value, Is.EqualTo(expectedValue));
+            });
         }
     }
 }